refactored tasks so that user tasks are rendered based on user.
